Added <SearchResults>.sift()
A small utility function in the SearchResults structure, to weed out posts with undesirable tags. A soft work around, to get through the limit of 6 tags on e6/e9, (and any others with tag limits)

If there's something I have to change or improve to fit the style, just let me know. I'm new to TypeScript.